Think i finished page fault. Prob not

+++ memory/memory_manager.h
@@ -12,7 +12,7 @@ int alloc_pt (int num_pages);
 int dealloc_pt (int page_table_index);
 int page_fault (int page_table_index, int page_num);
 //Returns either index of phy_mem or error code
-int lru_lookup();
+byte lru_lookup();
 short find_empty_back_addr();
 int set_back_addr_empty(short addr);
 int set_back_addr_full(short addr);
